src/requirement.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
docopt==0.6.2
Flask==2.2.5
flatbuffers==24.12.23
gast==0.4.0
google-auth==2.37.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.62.3
h5py==3.8.0
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.0.3
keras==2.11.0
libclang==18.1.1
Markdown==3.4.4
MarkupSafe==2.1.5
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==24.0
Pillow==9.5.0
pip==24.0
pkg_resources==0.0.0
protobuf==3.19.6
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
setuptools==68.0.0
six==1.17.0
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wi==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
wheel==0.42.0
wrapt==1.16.0
yarg==0.1.10
zipp==3.15.0
